refactor(api,db): 认证依赖从 db 包迁往 app/api/deps.py
鉴权是 HTTP 层的关注点，产出 403/400 而非数据。它住在 app/db/user_oper.py 里让数据层反向依赖 fastapi，也使鉴权无法与数据访问分开度量。

────────────────────────────────────────────────────────────────────────
以下为合并前各原始提交的完整说明（逐字保留其推理与踩坑记录）：
────────────────────────────────────────────────────────────────────────

原始提交 4b53729dd — refactor(api,db): 认证依赖从 db 包迁往 app/api/deps.py

user_oper.py 里有 9 个模块级函数是 FastAPI 路由依赖——从令牌解出用户、校验
激活状态与权限，产出的是 403/400 而不是数据。它们住在 db 包里造成两个问题：
数据层反向依赖 fastapi；鉴权逻辑与数据访问混在一个文件，覆盖率无法分开度量
（整体卡在 58%，未覆盖的全是需要构造 JWT 的那部分）。

迁至 app/api/deps.py——FastAPI 的惯例位置，且消费方本来就只有
app/api/endpoints/ 下的端点与授权测试。函数体逐字未改。

旧导入路径经模块级 __getattr__ 惰性转发保留：第三方插件在仓库外、不可枚举，
直接移除会让它们在 import 期崩。用惰性而非模块级 re-import，是为了避免
app.db.user_oper -> app.api.deps -> app.db 在导入期成环。

user_oper.py 191 -> 116 行；拆分后 user_oper 覆盖率 58.88% -> 75.41%，
迁出的鉴权部分单独计为 42.86%，两者可分别推进。全量 4055 passed。
