+++ app/api/endpoints/llm.py
@@ -15,10 +15,7 @@ from app.agent.llm import (
 )
 from app.runtime.config import settings
 from app.db.models import User
-from app.db.user_oper import (
-    get_current_active_superuser_async,
-    get_current_active_user_async,
-)
+from app.api.deps import get_current_active_superuser_async, get_current_active_user_async
 from app.runtime.log import logger
 
 router = ResponseAPIRouter()
