+++ app/api/endpoints/media.py
@@ -17,7 +17,7 @@ from app.domain.meta.metamusic import MetaMusic
 from app.domain.metainfo import MetaInfo, MetaInfoPath
 from app.application.security.access import verify_token, verify_apitoken
 from app.db.models import User
-from app.db.user_oper import get_current_active_user, get_current_active_superuser
+from app.api.deps import get_current_active_user, get_current_active_superuser
 from app.schemas import MediaType
 from app.schemas.category import CategoryConfig
 from app.schemas.types import MUSIC_ENTITY_RECORDING, MediaSource
