+++ app/api/endpoints/message.py
@@ -16,7 +16,7 @@ from app.db import get_async_db
 from app.db.models import User
 from app.db.message_oper import MessageOper
 from app.db.systemconfig_oper import SystemConfigOper
-from app.db.user_oper import get_current_active_superuser
+from app.api.deps import get_current_active_superuser
 from app.runtime.extensions.service_registry import ServiceConfigHelper
 from app.runtime.log import logger
 from app.modules.wechat.WXBizMsgCrypt3 import WXBizMsgCrypt
