+++ app/api/endpoints/music.py
@@ -10,7 +10,7 @@ from app.schemas.types import MediaSource, MediaType
 from app.domain.context import MusicAlbumInfo, MusicArtistInfo, MusicInfo
 from app.application.security.access import verify_token
 from app.db.models.user import User
-from app.db.user_oper import get_current_active_superuser_async
+from app.api.deps import get_current_active_superuser_async
 from app.modules.listenbrainz import (
     LISTENBRAINZ_CHART_RANGES,
     LISTENBRAINZ_FRESH_MAX_DAYS,
