+++ app/api/endpoints/notification.py
@@ -6,7 +6,7 @@ from app import schemas
 from app.api.response import ResponseAPIRouter
 from app.runtime.extensions.module_manager import ModuleManager
 from app.db.models import User
-from app.db.user_oper import get_current_active_superuser
+from app.api.deps import get_current_active_superuser
 from app.modules.wechatclawbot.wechatclawbot import WechatClawBot
 
 router = ResponseAPIRouter()
