+++ app/api/endpoints/plugin.py
@@ -25,10 +25,7 @@ from app.application.security.access import (
 )
 from app.db.models import User
 from app.db.systemconfig_oper import SystemConfigOper
-from app.db.user_oper import (
-    get_current_active_superuser,
-    get_current_active_superuser_async,
-)
+from app.api.deps import get_current_active_superuser, get_current_active_superuser_async
 from app.factory import app
 from app.adapters.external.server import MoviePilotServerHelper
 from app.adapters.external.market import PluginHelper
