+++ app/api/endpoints/system.py
@@ -30,11 +30,7 @@ from app.runtime.extensions.module_manager import ModuleManager
 from app.application.security.access import verify_apitoken, verify_resource_token, verify_token
 from app.db.models import User
 from app.db.systemconfig_oper import SystemConfigOper
-from app.db.user_oper import (
-    get_current_active_superuser,
-    get_current_active_superuser_async,
-    get_current_active_user_async,
-)
+from app.api.deps import get_current_active_superuser, get_current_active_superuser_async, get_current_active_user_async
 from app.application.image import ImageHelper
 from app.runtime.localization import LocaleHelper
 from app.adapters.external.market import (
