+++ app/api/endpoints/tmdb.py
@@ -9,7 +9,7 @@ from app.runtime.config import settings
 from app.application.security.access import verify_token
 from app.db.models.user import User
 from app.db.systemconfig_oper import SystemConfigOper
-from app.db.user_oper import get_current_active_superuser_async
+from app.api.deps import get_current_active_superuser_async
 from app.modules.themoviedb.tmdb_cache import TmdbCache
 from app.schemas.types import MediaType, SystemConfigKey
