+++ app/api/endpoints/transfer.py
@@ -13,10 +13,7 @@ from app.application.security.access import verify_token, verify_apitoken
 from app.db import get_db
 from app.db.models import User
 from app.db.models.transferhistory import TransferHistory
-from app.db.user_oper import (
-    get_current_active_manage_user,
-    get_current_active_superuser,
-)
+from app.api.deps import get_current_active_manage_user, get_current_active_superuser
 from app.application.directory import DirectoryHelper
 from app.runtime.log import logger
 from app.schemas import (
