+++ app/api/endpoints/workflow.py
@@ -15,10 +15,7 @@ from app.workflow import WorkFlowManager
 from app.db import get_async_db, get_db
 from app.db.models import Workflow, User
 from app.db.systemconfig_oper import SystemConfigOper
-from app.db.user_oper import (
-    get_current_active_manage_user,
-    get_current_active_manage_user_async,
-)
+from app.api.deps import get_current_active_manage_user, get_current_active_manage_user_async
 from app.db.workflow_oper import WorkflowOper
 from app.adapters.external.server import MoviePilotServerHelper
 from app.scheduler import Scheduler
