+++ tests/test_music_recognize_cache.py
@@ -11,7 +11,7 @@ from unittest.mock import Mock
 from app.api.endpoints import music as music_endpoint
 from app.domain.context import MusicInfo
 from app.domain.meta.metamusic import MetaMusic
-from app.db.user_oper import get_current_active_superuser_async
+from app.api.deps import get_current_active_superuser_async
 from app.modules.musicbrainz import music_cache as music_cache_module
 from app.modules.musicbrainz import MusicBrainzModule
 from app.modules.musicbrainz.music_cache import MusicBrainzCache
