+++ tests/test_workflow_authorization.py
@@ -8,7 +8,7 @@ from fastapi.routing import APIRoute
 
 from app.api.endpoints import workflow as workflow_endpoint
 from app.application.security.access import verify_token
-from app.db.user_oper import (
+from app.api.deps import (
     get_current_active_manage_user,
     get_current_active_manage_user_async,
     get_current_active_user,
